refactor(composite): extract aggregation into separate modules
- Extract AggregationConfig, AggregationFieldSpec, AggregationFunction,
  EnricherCardinality from config.py to aggregation.py (domain)
- Create EnricherAggregator class in aggregator.py (application)
- Delegate aggregation from MergeService to EnricherAggregator
- Extract _resolve_join_key_names helper to reduce _apply_joins complexity
- Update tests to use new aggregator fixture
- Fix architecture metrics violations:
  * config.py: 649 -> 627 LOC (limit: 640)
  * merger.py: _apply_joins CC=17 -> CC<15
