+++ tests/unit/application/composite/test_merger.py
@@ -6,16 +6,10 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
+from bioetl.application.composite.aggregator import EnricherAggregator
 from bioetl.application.composite.deduplication import EnricherDeduplicator
 from bioetl.application.composite.merger import MergeService, _path_to_table_name
-from bioetl.domain.composite.config import (
-    AggregationConfig,
-    AggregationFieldSpec,
-    AggregationFunction,
-    EnricherCardinality,
-    EnricherConfig,
-    MergeConfig,
-)
+from bioetl.domain.composite.config import EnricherConfig, MergeConfig
 from bioetl.domain.composite.result import EnrichmentResult, EnrichmentStatus
 from bioetl.domain.composite.strategy import ConflictResolution, MergeStrategy
 
@@ -40,6 +34,12 @@ def mock_logger():
 def deduplicator(mock_logger):
     """Create an EnricherDeduplicator instance."""
     return EnricherDeduplicator(mock_logger)
+
+
+@pytest.fixture
+def aggregator(mock_logger):
+    """Create an EnricherAggregator instance."""
+    return EnricherAggregator(mock_logger)
 
 
 @pytest.fixture
@@ -1546,7 +1546,7 @@ class TestQualifiedJoinKeys:
 class TestManyToOneAggregation:
     """Tests for 1:M enricher aggregation."""
 
-    def test_aggregate_collect_list(self, merge_service):
+    def test_aggregate_collect_list(self, aggregator):
         """Test COLLECT_LIST aggregation function."""
         import polars as pl
         from bioetl.domain.composite.config import (
@@ -1555,11 +1555,13 @@ class TestManyToOneAggregation:
             AggregationFunction,
         )
 
-        df = pl.DataFrame({
-            "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL2"],
-            "term": ["Aspirin", "Pain", "Kinase"],
-            "term_type": ["MESH_HEADING", "MESH_HEADING", "KEYWORD"],
-        })
+        df = pl.DataFrame(
+            {
+                "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL2"],
+                "term": ["Aspirin", "Pain", "Kinase"],
+                "term_type": ["MESH_HEADING", "MESH_HEADING", "KEYWORD"],
+            }
+        )
 
         config = AggregationConfig(
             group_by="document_chembl_id",
@@ -1572,17 +1574,17 @@ class TestManyToOneAggregation:
             ),
         )
 
-        result = merge_service._aggregate_enricher_data(df, config, "test_enricher")
+        result = aggregator.aggregate(df, config, "test_enricher")
 
         assert len(result) == 2
         assert set(result["document_chembl_id"].to_list()) == {"CHEMBL1", "CHEMBL2"}
 
-        chembl1_terms = result.filter(
-            pl.col("document_chembl_id") == "CHEMBL1"
-        )["all_terms"][0]
+        chembl1_terms = result.filter(pl.col("document_chembl_id") == "CHEMBL1")[
+            "all_terms"
+        ][0]
         assert set(chembl1_terms) == {"Aspirin", "Pain"}
 
-    def test_aggregate_collect_set(self, merge_service):
+    def test_aggregate_collect_set(self, aggregator):
         """Test COLLECT_SET aggregation function (unique values)."""
         import polars as pl
         from bioetl.domain.composite.config import (
@@ -1591,10 +1593,12 @@ class TestManyToOneAggregation:
             AggregationFunction,
         )
 
-        df = pl.DataFrame({
-            "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL1"],
-            "term": ["Aspirin", "Aspirin", "Pain"],  # Duplicate Aspirin
-        })
+        df = pl.DataFrame(
+            {
+                "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL1"],
+                "term": ["Aspirin", "Aspirin", "Pain"],  # Duplicate Aspirin
+            }
+        )
 
         config = AggregationConfig(
             group_by="document_chembl_id",
@@ -1607,7 +1611,7 @@ class TestManyToOneAggregation:
             ),
         )
 
-        result = merge_service._aggregate_enricher_data(df, config, "test_enricher")
+        result = aggregator.aggregate(df, config, "test_enricher")
 
         assert len(result) == 1
         terms = result["unique_terms"][0]
@@ -1615,7 +1619,7 @@ class TestManyToOneAggregation:
         assert len(terms) == 2
         assert set(terms) == {"Aspirin", "Pain"}
 
-    def test_aggregate_count(self, merge_service):
+    def test_aggregate_count(self, aggregator):
         """Test COUNT aggregation function."""
         import polars as pl
         from bioetl.domain.composite.config import (
@@ -1624,10 +1628,12 @@ class TestManyToOneAggregation:
             AggregationFunction,
         )
 
-        df = pl.DataFrame({
-            "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL1", "CHEMBL2"],
-            "term": ["A", "B", "C", "D"],
-        })
+        df = pl.DataFrame(
+            {
+                "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL1", "CHEMBL2"],
+                "term": ["A", "B", "C", "D"],
+            }
+        )
 
         config = AggregationConfig(
             group_by="document_chembl_id",
@@ -1640,20 +1646,20 @@ class TestManyToOneAggregation:
             ),
         )
 
-        result = merge_service._aggregate_enricher_data(df, config, "test_enricher")
+        result = aggregator.aggregate(df, config, "test_enricher")
 
         assert len(result) == 2
-        chembl1_count = result.filter(
-            pl.col("document_chembl_id") == "CHEMBL1"
-        )["term_count"][0]
+        chembl1_count = result.filter(pl.col("document_chembl_id") == "CHEMBL1")[
+            "term_count"
+        ][0]
         assert chembl1_count == 3
 
-        chembl2_count = result.filter(
-            pl.col("document_chembl_id") == "CHEMBL2"
-        )["term_count"][0]
+        chembl2_count = result.filter(pl.col("document_chembl_id") == "CHEMBL2")[
+            "term_count"
+        ][0]
         assert chembl2_count == 1
 
-    def test_aggregate_first(self, merge_service):
+    def test_aggregate_first(self, aggregator):
         """Test FIRST aggregation function."""
         import polars as pl
         from bioetl.domain.composite.config import (
@@ -1662,10 +1668,12 @@ class TestManyToOneAggregation:
             AggregationFunction,
         )
 
-        df = pl.DataFrame({
-            "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL1"],
-            "term": ["First", "Second", "Third"],
-        })
+        df = pl.DataFrame(
+            {
+                "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL1"],
+                "term": ["First", "Second", "Third"],
+            }
+        )
 
         config = AggregationConfig(
             group_by="document_chembl_id",
@@ -1678,12 +1686,12 @@ class TestManyToOneAggregation:
             ),
         )
 
-        result = merge_service._aggregate_enricher_data(df, config, "test_enricher")
+        result = aggregator.aggregate(df, config, "test_enricher")
 
         assert len(result) == 1
         assert result["first_term"][0] == "First"
 
-    def test_aggregate_concat_str(self, merge_service):
+    def test_aggregate_concat_str(self, aggregator):
         """Test CONCAT_STR aggregation function."""
         import polars as pl
         from bioetl.domain.composite.config import (
@@ -1692,10 +1700,12 @@ class TestManyToOneAggregation:
             AggregationFunction,
         )
 
-        df = pl.DataFrame({
-            "document_chembl_id": ["CHEMBL1", "CHEMBL1"],
-            "term": ["Aspirin", "Pain"],
-        })
+        df = pl.DataFrame(
+            {
+                "document_chembl_id": ["CHEMBL1", "CHEMBL1"],
+                "term": ["Aspirin", "Pain"],
+            }
+        )
 
         config = AggregationConfig(
             group_by="document_chembl_id",
@@ -1708,12 +1718,12 @@ class TestManyToOneAggregation:
             ),
         )
 
-        result = merge_service._aggregate_enricher_data(df, config, "test_enricher")
+        result = aggregator.aggregate(df, config, "test_enricher")
 
         assert len(result) == 1
         assert result["terms_str"][0] == "Aspirin, Pain"
 
-    def test_aggregate_with_filter(self, merge_service):
+    def test_aggregate_with_filter(self, aggregator):
         """Test aggregation with filter condition."""
         import polars as pl
         from bioetl.domain.composite.config import (
@@ -1722,11 +1732,13 @@ class TestManyToOneAggregation:
             AggregationFunction,
         )
 
-        df = pl.DataFrame({
-            "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL1"],
-            "term": ["Aspirin", "Pain", "Kinase"],
-            "term_type": ["MESH_HEADING", "MESH_HEADING", "KEYWORD"],
-        })
+        df = pl.DataFrame(
+            {
+                "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL1"],
+                "term": ["Aspirin", "Pain", "Kinase"],
+                "term_type": ["MESH_HEADING", "MESH_HEADING", "KEYWORD"],
+            }
+        )
 
         config = AggregationConfig(
             group_by="document_chembl_id",
@@ -1740,13 +1752,13 @@ class TestManyToOneAggregation:
             ),
         )
 
-        result = merge_service._aggregate_enricher_data(df, config, "test_enricher")
+        result = aggregator.aggregate(df, config, "test_enricher")
 
         mesh_terms = result["mesh_terms"][0]
         assert set(mesh_terms) == {"Aspirin", "Pain"}
         assert "Kinase" not in mesh_terms
 
-    def test_aggregate_multiple_fields(self, merge_service):
+    def test_aggregate_multiple_fields(self, aggregator):
         """Test aggregation with multiple fields."""
         import polars as pl
         from bioetl.domain.composite.config import (
@@ -1755,11 +1767,13 @@ class TestManyToOneAggregation:
             AggregationFunction,
         )
 
-        df = pl.DataFrame({
-            "document_chembl_id": ["CHEMBL1", "CHEMBL1"],
-            "term": ["Aspirin", "Pain"],
-            "mesh_id": ["D001", "D002"],
-        })
+        df = pl.DataFrame(
+            {
+                "document_chembl_id": ["CHEMBL1", "CHEMBL1"],
+                "term": ["Aspirin", "Pain"],
+                "mesh_id": ["D001", "D002"],
+            }
+        )
 
         config = AggregationConfig(
             group_by="document_chembl_id",
@@ -1782,14 +1796,14 @@ class TestManyToOneAggregation:
             ),
         )
 
-        result = merge_service._aggregate_enricher_data(df, config, "test_enricher")
+        result = aggregator.aggregate(df, config, "test_enricher")
 
         assert len(result) == 1
         assert set(result["terms"][0]) == {"Aspirin", "Pain"}
         assert set(result["mesh_ids"][0]) == {"D001", "D002"}
         assert result["term_count"][0] == 2
 
-    def test_aggregate_with_null_values(self, merge_service):
+    def test_aggregate_with_null_values(self, aggregator):
         """Test aggregation handles null values correctly."""
         import polars as pl
         from bioetl.domain.composite.config import (
@@ -1798,10 +1812,12 @@ class TestManyToOneAggregation:
             AggregationFunction,
         )
 
-        df = pl.DataFrame({
-            "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL1"],
-            "term": ["Aspirin", None, "Pain"],
-        })
+        df = pl.DataFrame(
+            {
+                "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL1"],
+                "term": ["Aspirin", None, "Pain"],
+            }
+        )
 
         config = AggregationConfig(
             group_by="document_chembl_id",
@@ -1814,18 +1830,18 @@ class TestManyToOneAggregation:
             ),
         )
 
-        result = merge_service._aggregate_enricher_data(df, config, "test_enricher")
+        result = aggregator.aggregate(df, config, "test_enricher")
 
         # Nulls should be dropped
         terms = result["terms"][0]
         assert len(terms) == 2
         assert set(terms) == {"Aspirin", "Pain"}
 
-    def test_parse_filter_is_not_null(self, merge_service):
+    def test_parse_filter_is_not_null(self, aggregator):
         """Test parsing IS NOT NULL filter condition."""
         import polars as pl
 
-        expr = merge_service._parse_filter_condition("field IS NOT NULL")
+        expr = aggregator._parse_filter_condition("field IS NOT NULL")
         assert expr is not None
 
         # Test the expression works
@@ -1833,11 +1849,11 @@ class TestManyToOneAggregation:
         result = df.filter(expr)
         assert len(result) == 2
 
-    def test_parse_filter_is_null(self, merge_service):
+    def test_parse_filter_is_null(self, aggregator):
         """Test parsing IS NULL filter condition."""
         import polars as pl
 
-        expr = merge_service._parse_filter_condition("field IS NULL")
+        expr = aggregator._parse_filter_condition("field IS NULL")
         assert expr is not None
 
         df = pl.DataFrame({"field": ["a", None, "b"]})
@@ -1845,31 +1861,31 @@ class TestManyToOneAggregation:
         assert len(result) == 1
         assert result["field"][0] is None
 
-    def test_parse_filter_equality(self, merge_service):
+    def test_parse_filter_equality(self, aggregator):
         """Test parsing equality filter condition."""
         import polars as pl
 
-        expr = merge_service._parse_filter_condition("term_type == 'MESH_HEADING'")
+        expr = aggregator._parse_filter_condition("term_type == 'MESH_HEADING'")
         assert expr is not None
 
         df = pl.DataFrame({"term_type": ["MESH_HEADING", "KEYWORD", "MESH_HEADING"]})
         result = df.filter(expr)
         assert len(result) == 2
 
-    def test_parse_filter_inequality(self, merge_service):
+    def test_parse_filter_inequality(self, aggregator):
         """Test parsing inequality filter condition."""
         import polars as pl
 
-        expr = merge_service._parse_filter_condition("term_type != 'KEYWORD'")
+        expr = aggregator._parse_filter_condition("term_type != 'KEYWORD'")
         assert expr is not None
 
         df = pl.DataFrame({"term_type": ["MESH_HEADING", "KEYWORD", "CONCEPT"]})
         result = df.filter(expr)
         assert len(result) == 2
 
-    def test_parse_filter_invalid_returns_none(self, merge_service):
+    def test_parse_filter_invalid_returns_none(self, aggregator):
         """Test invalid filter returns None."""
-        expr = merge_service._parse_filter_condition("some random text")
+        expr = aggregator._parse_filter_condition("some random text")
         assert expr is None
 
     @pytest.mark.asyncio
@@ -1883,17 +1899,21 @@ class TestManyToOneAggregation:
             EnricherCardinality,
         )
 
-        seed_df = pl.DataFrame({
-            "document_chembl_id": ["CHEMBL1", "CHEMBL2"],
-            "title": ["Study A", "Study B"],
-        })
+        seed_df = pl.DataFrame(
+            {
+                "document_chembl_id": ["CHEMBL1", "CHEMBL2"],
+                "title": ["Study A", "Study B"],
+            }
+        )
 
         # Enricher has multiple rows per document
-        enricher_df = pl.DataFrame({
-            "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL1", "CHEMBL2"],
-            "term": ["Aspirin", "Pain", "Drug", "Kinase"],
-            "term_type": ["MESH_HEADING", "MESH_HEADING", "KEYWORD", "KEYWORD"],
-        })
+        enricher_df = pl.DataFrame(
+            {
+                "document_chembl_id": ["CHEMBL1", "CHEMBL1", "CHEMBL1", "CHEMBL2"],
+                "term": ["Aspirin", "Pain", "Drug", "Kinase"],
+                "term_type": ["MESH_HEADING", "MESH_HEADING", "KEYWORD", "KEYWORD"],
+            }
+        )
 
         enricher_config = EnricherConfig(
             pipeline="chembl_publication_term",
@@ -1931,10 +1951,10 @@ class TestManyToOneAggregation:
         # Check aggregated values
         chembl1 = result.filter(pl.col("document_chembl_id") == "CHEMBL1")
         # mesh_headings column should exist with qualified name
-        mesh_col = [c for c in result.columns if "mesh_headings" in c][0]
+        mesh_col = next(c for c in result.columns if "mesh_headings" in c)
         mesh_terms = chembl1[mesh_col][0]
         assert set(mesh_terms) == {"Aspirin", "Pain"}
 
         # term_count should include all terms
-        count_col = [c for c in result.columns if "term_count" in c][0]
+        count_col = next(c for c in result.columns if "term_count" in c)
         assert chembl1[count_col][0] == 3  # Aspirin, Pain, Drug
